sort strassen:
  move useful generic bits to AA
  useful special bits to Nemo
  Det code to Det.jl
